requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
langchain==0.2.11
langchain-google-genai==1.0.7
langchain-core==0.2.23
python-dotenv==1.0.0
typing-extensions==4.8.0
playwright
mangum
